+++ PyDodo/pydodo/create_aircraft.py
@@ -1,6 +1,6 @@
 import requests
 
-from . import utils
+from . import validate_input
 from .utils import post_request
 from .config_param import config_param
 
@@ -18,16 +18,16 @@ def create_aircraft(
     """
 	Create new aircraft. Raises error if inputs are invalid or if aircraft already exists.
 	"""
-    utils._check_string_input(aircraft_id, "aircraft_id")
-    utils._check_string_input(type, "type")
-    utils._check_latitude(latitude)
-    utils._check_longitude(longitude)
-    utils._check_heading(heading)
-    utils._check_speed(speed)
+    validate_input._check_type_string(aircraft_id, "aircraft_id")
+    validate_input._check_type_string(type, "type")
+    validate_input._check_latitude(latitude)
+    validate_input._check_longitude(longitude)
+    validate_input._check_heading(heading)
+    validate_input._check_speed(speed)
     assert (
         altitude is None or flight_level is None
     ), "Only altitude or flight level should be provided, not both"
-    alt = utils.parse_alt(altitude, flight_level)
+    alt = validate_input.parse_alt(altitude, flight_level)
 
     json = {
         "acid": aircraft_id,
